Rename tests/utils_for_test.py to tests/utils.py
